refactor: standardize legacy module names
