+++ src/mindroom/config/legacy_access.py
