+++ src/mindroom/event_journal/legacy_schema.py
