+++ src/mindroom/mcp_gateway/legacy_schema.py
