+++ src/mindroom/oauth/legacy_credentials.py
