+++ src/mindroom/script_runs/legacy_schema.py
