+++ tests/test_private_instance_identity_store.py
@@ -8,8 +8,8 @@ from typing import TYPE_CHECKING
 
 import pytest
 
+from mindroom import legacy_private_storage_aliases as compat
 from mindroom import private_instance_identity_store as store
-from mindroom import private_storage_compat as compat
 from mindroom.tool_system.worker_routing import private_instance_scope_root_path
 
 if TYPE_CHECKING:
