+++ tests/test_private_storage_migration.py
@@ -106,7 +106,7 @@ async def test_startup_moves_every_owner_and_preserves_contents(tmp_path: Path, 
     ]
     primary = [_files(source) for source in sources]
     secondary = [_files(resolve_session_state_root(source, paths)) for source in sources]
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     await migration.migrate_private_storage(paths)
     for index, (_old, requester, current) in enumerate(fixtures):
         target = private_instance_scope_root_path(paths.storage_root, current)
@@ -131,7 +131,7 @@ _REQUESTER = "@alice:example.org"
 @pytest.mark.asyncio
 async def test_every_filesystem_mutation_boundary_resumes(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
     """Crashes around actual fsync, replace, rename and unlink never lose scope contents."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     observed: list[str] = []
     originals = {name: getattr(os, name) for name in ("fsync", "replace", "rename", "unlink", "symlink")}
 
@@ -192,7 +192,7 @@ async def test_preflight_rejects_conflicts_before_any_move(
     conflict: str,
 ) -> None:
     """A conflicting later scope must not leave an earlier verified scope partially moved."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     source = _seed(paths, _OLD, _REQUESTER)
     target = private_instance_scope_root_path(paths.storage_root, _NEW)
@@ -233,7 +233,7 @@ async def test_current_and_fresh_startup_skip_workers_and_contents(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
     """Legitimate new traffic never causes a completed scope to be migrated again."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     preflight = importlib.import_module("mindroom.workers.storage_preflight")
     paths = _paths(tmp_path)
     _seed(paths, _OLD, _REQUESTER)
@@ -256,7 +256,7 @@ async def test_current_and_fresh_startup_skip_workers_and_contents(
 
 
 async def _interrupt_after_session_move(paths: RuntimePaths, monkeypatch: pytest.MonkeyPatch) -> Path:
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     rename = Path.rename
 
     def interrupt(source: Path, target: Path) -> Path:
@@ -282,7 +282,7 @@ async def test_recovery_rejects_missing_or_unrelated_evidence(
     damage: str,
 ) -> None:
     """An intent authorizes only its exact roots, original directories and saved owner."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     _seed(paths, _OLD, _REQUESTER)
     source = await _interrupt_after_session_move(paths, monkeypatch)
@@ -318,7 +318,7 @@ async def test_recovery_rejects_missing_or_unrelated_evidence(
 @pytest.mark.asyncio
 async def test_recovery_accepts_remount_with_same_inodes(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
     """Kernel device identifiers may change across mounts without invalidating durable intent."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     _seed(paths, _OLD, _REQUESTER)
     source = await _interrupt_after_session_move(paths, monkeypatch)
@@ -347,7 +347,7 @@ async def test_recovery_accepts_remount_with_same_inodes(tmp_path: Path, monkeyp
 @pytest.mark.parametrize("kind", ["absolute_old", "absolute_new", "relative_escape", "safe_absolute", "safe_relative"])
 async def test_relocation_checks_links_without_rewriting_them(tmp_path: Path, kind: str) -> None:
     """Historical aliases keep opaque link targets unchanged, including absolute references."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     source = _seed(paths, _OLD, _REQUESTER)
     link = source / "link"
@@ -370,7 +370,7 @@ async def test_cancellation_keeps_locks_until_blocking_migration_drains(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
     """Cancelled admission cannot release volume locks while the worker thread still moves state."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     preflight = importlib.import_module("mindroom.workers.storage_preflight")
     paths = _paths(tmp_path)
     source = _seed(paths, _OLD, _REQUESTER)
@@ -424,7 +424,7 @@ async def test_primary_admission_fails_before_credentials(
 @pytest.mark.parametrize("payload", ["null", "false", "[]", '{"version": 1}'])
 async def test_malformed_intent_never_becomes_fresh_migration(tmp_path: Path, payload: str) -> None:
     """A present but invalid intent must not be replaced with newly inferred evidence."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     source = _seed(paths, _OLD, _REQUESTER)
     (source / _INTENT).write_text(payload)
@@ -441,7 +441,7 @@ async def test_nested_mounts_fail_before_any_move(
     location: str,
 ) -> None:
     """Scope, nested-directory and namespace mount boundaries cannot be moved by startup."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     source = _seed(paths, _OLD, _REQUESTER)
     mounted = {
@@ -470,7 +470,7 @@ async def test_nested_mounts_fail_before_any_move(
 @pytest.mark.asyncio
 async def test_previously_colliding_requesters_keep_distinct_state(tmp_path: Path) -> None:
     """A colon-bearing owner and its underscore lookalike remain isolated after startup."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     first = _seed(paths, _OLD, _REQUESTER)
     second = _seed(paths, "v1:default:user:@alice_example.org", "@alice_example.org")
@@ -488,7 +488,7 @@ async def test_previously_colliding_requesters_keep_distinct_state(tmp_path: Pat
 @pytest.mark.asyncio
 async def test_missing_configured_session_volume_blocks_startup(tmp_path: Path) -> None:
     """A typo or missing mounted session volume must not silently discard the mirror."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     source = _seed(paths, _OLD, _REQUESTER)
     paths = replace(paths, process_env={"MINDROOM_SESSION_STORAGE_PATH": str(tmp_path / "missing")})
@@ -500,7 +500,7 @@ async def test_missing_configured_session_volume_blocks_startup(tmp_path: Path) 
 @pytest.mark.asyncio
 async def test_absent_optional_mirror_stays_absent_on_resume(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
     """An explicitly absent mirror cannot acquire unrelated data during recovery."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     source = _seed(paths, _OLD, _REQUESTER)
     shutil.rmtree(resolve_session_state_root(source, paths))
@@ -518,7 +518,7 @@ async def test_worker_preflight_failure_preserves_original_layout(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
     """Storage is untouched when managed workers cannot be proven absent."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     preflight = importlib.import_module("mindroom.workers.storage_preflight")
     paths = _paths(tmp_path)
     source = _seed(paths, _OLD, _REQUESTER)
@@ -539,7 +539,7 @@ async def test_worker_preflight_failure_preserves_original_layout(
 @pytest.mark.asyncio
 async def test_recovery_rejects_boolean_owner_version(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
     """Intent recovery must retain the strict owner schema used for fresh discovery."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     _seed(paths, _OLD, _REQUESTER)
     source = await _interrupt_after_session_move(paths, monkeypatch)
@@ -554,7 +554,7 @@ async def test_recovery_rejects_boolean_owner_version(tmp_path: Path, monkeypatc
 @pytest.mark.asyncio
 async def test_unreadable_pending_tree_stops_batch(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
     """The non-Linux mount scan fails closed when directory metadata is unreadable."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     source = _seed(paths, _OLD, _REQUESTER)
     monkeypatch.setattr(migration.sys, "platform", "darwin")
@@ -579,7 +579,7 @@ async def test_abrupt_process_exit_leaves_harmless_partial_temporary(
     reserved: str,
 ) -> None:
     """Real process death bypasses durable-writer cleanup; partial temporaries stay opaque on retry."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     source = _seed(paths, _OLD, _REQUESTER)
     program = textwrap.dedent("""
@@ -589,7 +589,7 @@ async def test_abrupt_process_exit_leaves_harmless_partial_temporary(
         import sys
         from pathlib import Path
         from mindroom.constants import resolve_runtime_paths
-        from mindroom.private_storage_migration import migrate_private_storage
+        from mindroom.legacy_private_storage import migrate_private_storage
         paths = resolve_runtime_paths(
             config_path=Path(sys.argv[1]), storage_path=Path(sys.argv[2]),
             process_env={"MINDROOM_SESSION_STORAGE_PATH": sys.argv[3]},
@@ -637,7 +637,7 @@ async def test_abrupt_process_exit_leaves_harmless_partial_temporary(
 @pytest.mark.parametrize("scope", ["user", "user_agent"])
 async def test_current_runtime_export_and_mounts_find_migrated_contents(tmp_path: Path, scope: str) -> None:
     """Normal consumers resolve moved owner data through unchanged current-key paths."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     old_key = _OLD if scope == "user" else "v1:default:user_agent:@alice:example.org:writer"
     source = _seed(paths, old_key, _REQUESTER)
@@ -718,7 +718,7 @@ async def test_retry_durably_publishes_existing_intent_before_any_move(
     retry_sync_fails: bool,
 ) -> None:
     """A visible intent left by failed directory fsync must become durable before recovery renames."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     source = _seed(paths, _OLD, _REQUESTER)
     mirror = resolve_session_state_root(source, paths)
@@ -777,7 +777,7 @@ async def test_retry_durably_publishes_existing_intent_before_any_move(
 @pytest.mark.parametrize("indirect", [False, True])
 async def test_relative_link_chain_keeps_old_absolute_paths(tmp_path: Path, indirect: bool) -> None:
     """Links through external directories can still return to retained historical names."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     source = _seed(paths, _OLD, _REQUESTER)
     other = _seed(paths, "v1:default:user:@bob:example.org", "@bob:example.org")
@@ -802,7 +802,7 @@ async def test_relative_link_chain_keeps_old_absolute_paths(tmp_path: Path, indi
 @pytest.mark.parametrize("kind", ["contained_chain", "dangling", "cycle"])
 async def test_relative_link_chain_policy(tmp_path: Path, kind: str) -> None:
     """Contained, dangling and cyclic workspace links remain opaque and unchanged."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     source = _seed(paths, _OLD, _REQUESTER)
     targets = {"contained_chain": "writer/workspace", "dangling": "missing", "cycle": "link"}
@@ -820,7 +820,7 @@ async def test_relative_link_chain_policy(tmp_path: Path, kind: str) -> None:
 @pytest.mark.parametrize("separate", [False, True])
 async def test_historical_absolute_shebang_survives_migration(tmp_path: Path, separate: bool) -> None:
     """Executable bytes and both absolute names remain usable after relocation."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path, separate=separate)
     old = _seed(paths, _OLD, _REQUESTER)
     command = old / "writer/workspace/example-cli"
@@ -858,7 +858,7 @@ async def test_linux_bind_mounts_fail_before_moves(
     escaped: bool,
 ) -> None:
     """The mount table detects same-device binds, including kernel-escaped mountpoints."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     base = tmp_path / "space tab\t slash\\" if escaped else tmp_path
     base.mkdir(exist_ok=True)
     paths = _paths(base)
@@ -897,7 +897,7 @@ async def test_linux_unavailable_mount_table_fails_closed(
     table: str | None,
 ) -> None:
     """Missing or malformed Linux mount evidence cannot authorize a rename."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     source = _seed(paths, _OLD, _REQUESTER)
     original = Path.read_text
@@ -924,7 +924,7 @@ async def test_linux_reads_mount_table_once_without_walking_workspaces(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
     """Configured volume mounts are allowed and preflight work is bounded by mount count."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     source = _seed(paths, _OLD, _REQUESTER)
     roots = [paths.storage_root, resolve_session_state_root(paths.storage_root, paths)]
@@ -953,7 +953,7 @@ async def test_linux_reads_mount_table_once_without_walking_workspaces(
 @pytest.mark.parametrize("damage", ["foreign", "missing", "chain", "absolute", "owner", "mirror", "orphan_mirror"])
 async def test_completed_aliases_reject_tampering(tmp_path: Path, damage: str) -> None:
     """Completed aliases are accepted only at their exact owner-bound primary and mirror names."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     old = _seed(paths, _OLD, _REQUESTER)
     old_mirror = resolve_session_state_root(old, paths)
@@ -986,7 +986,7 @@ async def test_completed_aliases_reject_tampering(tmp_path: Path, damage: str) -
 @pytest.mark.asyncio
 async def test_current_scope_without_alias_is_not_given_historical_provenance(tmp_path: Path) -> None:
     """A current owner record alone cannot authorize creating an old path."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     current = _seed(paths, _NEW, _REQUESTER)
     await migration.migrate_private_storage(paths)
@@ -998,7 +998,7 @@ async def test_current_scope_without_alias_is_not_given_historical_provenance(tm
 @pytest.mark.asyncio
 async def test_empty_orphan_session_placeholder_remains_tolerated(tmp_path: Path) -> None:
     """Empty real placeholders keep prior startup tolerance without gaining alias authority."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     old = _seed(paths, _OLD, _REQUESTER)
     mirror = resolve_session_state_root(old, paths)
@@ -1016,7 +1016,7 @@ async def test_empty_orphan_session_placeholder_remains_tolerated(tmp_path: Path
 @pytest.mark.asyncio
 async def test_later_session_data_gets_no_guessed_alias(tmp_path: Path) -> None:
     """A mirror absent at migration may later hold legitimate current-key data."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     old = _seed(paths, _OLD, _REQUESTER)
     old_mirror = resolve_session_state_root(old, paths)
@@ -1044,7 +1044,7 @@ async def test_recovery_aliases_require_recorded_publication_state(
     damage: str,
 ) -> None:
     """An exact link target cannot excuse a wrong owner or impossible publication order."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     old = _seed(paths, _OLD, _REQUESTER)
     old_mirror = resolve_session_state_root(old, paths)
@@ -1080,7 +1080,7 @@ async def test_recovery_aliases_require_recorded_publication_state(
 @pytest.mark.asyncio
 async def test_alias_publication_does_not_overwrite_new_entry(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
     """A namespace conflict appearing after preflight remains intact with recovery evidence."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path, separate=False)
     old = _seed(paths, _OLD, _REQUESTER)
     original = os.symlink
@@ -1107,7 +1107,7 @@ async def test_separate_volumes_may_use_distinct_devices(
     platform: str,
 ) -> None:
     """Each scope rename stays on its own device; workspace symlinks are never traversed."""
-    migration = importlib.import_module("mindroom.private_storage_migration")
+    migration = importlib.import_module("mindroom.legacy_private_storage")
     paths = _paths(tmp_path)
     old = _seed(paths, _OLD, _REQUESTER)
     old_mirror = resolve_session_state_root(old, paths)
